[2/2] OmniSwitch: use old getTaskThumbnail for getting top activity
new snapshots are not updated for the current top app
which always leads to a empty thumb for newly started
apps. also fall back to the old method for any other
empty snapshot

and fix that stupid derp that causes reloading even
the preloaded thumbs. with that remove the useless async
loading setting too

Change-Id: I6301ea7bd05f581cc8d55f500d784716f0a4fd06